refactor: address review feedback (layer, dedup, comments)
- Move isBot from `lib/` to `infrastructure/http/` (next/server is server-runtime, not a UI utility)
- Extract `BotBlockedError` to shared `components/symbol-page/BotBlockedError.ts` and import in 3 useQuery-based hooks
- Remove decorative `// ---` separator comments in 3 Server Action test files
- Rewrite the `miss_no_trigger` branch comment in `useAnalysis.ts` to explain WHY a boolean flag is used instead of a discriminated union (useMutation cannot narrow on thrown error subtypes like useQuery)
