+++ src/__tests__/infrastructure/market/submitFundamentalAnalysisAction.test.ts
@@ -1,6 +1,4 @@
-// ---------------------------------------------------------------------------
 // Module mocks
-// ---------------------------------------------------------------------------
 
 jest.mock('@vercel/functions', () => ({
     waitUntil: jest.fn(),
@@ -31,9 +29,7 @@ jest.mock('@/infrastructure/market/byokGate', () => ({
     })),
 }));
 
-// ---------------------------------------------------------------------------
 // Typed mocks & fixtures
-// ---------------------------------------------------------------------------
 
 import {
     submitFundamentalAnalysis,
@@ -75,9 +71,7 @@ const gateError: AnalysisGateError = {
     message: 'mock-tier_premium_blocked',
 };
 
-// ---------------------------------------------------------------------------
 // Tests
-// ---------------------------------------------------------------------------
 
 describe('submitFundamentalAnalysisAction 함수는', () => {
     beforeEach(() => {
