+++ src/infrastructure/market/submitNewsAnalysisAction.ts
@@ -21,7 +21,7 @@ import {
     resolveTierAndByok,
     buildGateError,
 } from '@/infrastructure/market/byokGate';
-import { isBot } from '@/lib/isBot';
+import { isBot } from '@/infrastructure/http/isBot';
 import type { AnalysisGateBlockedResult } from '@/domain/types';
 
 /** Final return type — core's news result + our siglens-side gate errors. */
