+++ src/infrastructure/market/submitOverallAnalysisAction.ts
@@ -23,7 +23,7 @@ import {
     resolveTierAndByok,
     buildGateError,
 } from '@/infrastructure/market/byokGate';
-import { isBot } from '@/lib/isBot';
+import { isBot } from '@/infrastructure/http/isBot';
 import type { AnalysisGateBlockedResult } from '@/domain/types';
 
 /** Final return type — core's overall result + our siglens-side gate errors. */
